==> usrightmedia/code/08-topics/01-topic-model-inputs.ipynb <==
## Topic model inputs
- citation: based off of https://github.com/RaRe-Technologies/gensim/blob/develop/docs/notebooks/atmodel_tutorial.ipynb

**Purpose**: create corpus

import os
import pandas as pd
import pickle
from collections import defaultdict

# matplotlib is logged even though disable_existing_loggers=yes in logging_config.yaml
# https://stackoverflow.com/a/51529172/7016397
# workaround is to manually set the level before creating my logger
import logging
logging.getLogger('matplotlib').setLevel(logging.WARNING)

from usrightmedia.shared.loggers import get_logger
LOGGER = get_logger(filename = '01-topic-model-inputs', logger_type='main')

from usrightmedia.shared.topics_utils import *

### 1. Load documents from Elasticsearch

from inca import Inca
myinca = Inca()

query = {
    "query": {
        "bool": {
            "filter": [
                {"term": {"should_include": True}}
            ]
        }
    }
}

docs = myinca.database.document_generator(query)

titles = [] # docs option 1
leads = [] # docs option 2
texts = [] # docs option 3
doctypes = []
ids = []

for n, doc in enumerate(docs):
    title = doc["_source"]["title"]
    text = doc["_source"]["article_maintext"] # as scraped by news-please
    
    lead = text.split()[0:100]
    lead = ' '.join(lead) 
    lead = title + lead
    
    titles.append(title)
    leads.append(lead)
    texts.append(text)
    
    doctypes.append(doc["_source"]['doctype'])
    ids.append(doc['_id'])

df_data_inputs = pd.DataFrame({"doc_id": ids,
                               "doctype": doctypes,
                               "title": titles,
                               "lead": leads,
                               "article_maintext": texts})

df_data_inputs

%%time
with open(os.path.join(INPUTS_DIR, 'df_data_inputs.pkl'), 'wb') as file:
    pickle.dump(df_data_inputs, file)

### 2. Pre-process documents
- Removes empty documents to prepare for gensim training
- *Note*: ideally should have included `doc_id` during this pre-process step.
    - See `01b-topic-model-inputs-docs-with-ids.ipynb` for fix which allows topic assignments to be assigned to associated `doc_id`s.

- Manually install `en_core_web_lg` into virtual environment (https://github.com/explosion/spaCy/issues/4297)

```
conda activate usrightmedia

python -m spacy download en_core_web_lg
```

- not using trf (transformer-based model) because it has token limit of 512 (https://github.com/explosion/spaCy/issues/6939)


%%time
# empty docs removed
docs_titles = preprocess_docs(titles, 'titles', INPUTS_DIR)

%%time
# empty docs removed
docs_leads = preprocess_docs(leads, 'leads', INPUTS_DIR)

%%time
# empty docs removed
docs_texts = preprocess_docs(texts, 'texts', INPUTS_DIR)

### 3. Create dictionary

%%time
dict_titles = save_dictionary(docs_titles, 'titles', INPUTS_DIR)

%%time
dict_leads = save_dictionary(docs_leads, 'leads', INPUTS_DIR)

%%time
dict_texts = save_dictionary(docs_texts, 'texts', INPUTS_DIR)

### 4. Create corpus

%%time
corp_titles, corp_tfidf_titles = save_corpus(dict_titles, docs_titles, 'titles', INPUTS_DIR)

%%time
corp_leads, corp_tfidf_leads = save_corpus(dict_leads, docs_leads, 'leads', INPUTS_DIR)

%%time
corp_texts, corp_tfidf_texts = save_corpus(dict_texts, docs_texts, 'texts', INPUTS_DIR)

### 5. Summary

def print_input_summary(label, dictionary, corpus):
    print("-"*120)
    print(f"DOCS VERSION: {label}")
    print(f'Number of unique tokens: {len(dictionary)}')
    print(f'Number of documents: {len(corpus)}')

print_input_summary('titles', dict_titles, corp_titles)
print_input_summary('leads', dict_leads, corp_leads)
print_input_summary('texts', dict_texts, corp_texts)
