==> usrightmedia/code/08-topics/01b-topic-model-inputs-docs-with-ids.ipynb <==
## Topic model inputs with associated `doc_id`
- This step was done after the topic models were trained so that the topic assignments can be assigned back to the associated doc_ids.


### 1. Pre-process docs again
- *Note to self*: next time, don't forget to generate the gensim dictionary and corpus along with the associated IDs...

import os
import pandas as pd

from usrightmedia.shared.topics_utils import *

df_data_inputs = pd.read_pickle(os.path.join(INPUTS_DIR, 'df_data_inputs.pkl'))

df_data_inputs

%%time
# CPU times: user 16min 48s, sys: 635 ms, total: 16min 49s
# Wall time: 16min 49s
df_titles = preprocess_docs_with_doc_ids(df_data_inputs['doc_id'], df_data_inputs['title'], 'titles', INPUTS_DIR)

%%time
# CPU times: user 2h 16min 6s, sys: 23.5 s, total: 2h 16min 30s
# Wall time: 2h 16min 30s
df_leads = preprocess_docs_with_doc_ids(df_data_inputs['doc_id'], df_data_inputs['lead'], 'leads', INPUTS_DIR)

%%time
# CPU times: user 10h 47min 39s, sys: 1h 47min 22s, total: 12h 35min 2s
# Wall time: 12h 35min 20s
df_texts = preprocess_docs_with_doc_ids(df_data_inputs['doc_id'], df_data_inputs['article_maintext'], 'texts', INPUTS_DIR)

### 2. merge processed versions into `df_data_inputs`

#### 2.1 load files

df_titles = pd.read_pickle(os.path.join(INPUTS_DIR, "docs", "docs_titles_with_inca_ids.pkl"))
df_leads = pd.read_pickle(os.path.join(INPUTS_DIR, "docs", "docs_leads_with_inca_ids.pkl"))
df_texts = pd.read_pickle(os.path.join(INPUTS_DIR, "docs", "docs_texts_with_inca_ids.pkl"))

# not necessary since already loaded at the top of the notebook 
# df_data_inputs = pd.read_pickle(os.path.join(INPUTS_DIR, 'df_data_inputs.pkl'))

#### 2.2 rename columns to prep for merge

df_titles = df_titles.rename(columns={"processed_doc": "processed_title"})
df_leads = df_leads.rename(columns={"processed_doc": "processed_lead"})
df_texts = df_texts.rename(columns={"processed_doc": "processed_text"})

#### 2.3 preview dataframes which will be merged

df_titles

df_leads

df_texts

df_data_inputs

#### 2.4 merge dataframes

df_data_inputs = df_data_inputs.merge(right=df_titles,
                                      on="doc_id",
                                      how="left",
                                      validate="one_to_one")

df_data_inputs = df_data_inputs.merge(right=df_leads,
                                      on="doc_id",
                                      how="left",
                                      validate="one_to_one")

df_data_inputs = df_data_inputs.merge(right=df_texts,
                                      on="doc_id",
                                      how="left",
                                      validate="one_to_one")

df_data_inputs

%%time
with open(os.path.join(INPUTS_DIR, 'df_data_inputs_with_processed_docs.pkl'), 'wb') as file:
    pickle.dump(df_data_inputs, file)
