==> usrightmedia/code/08-topics/06-add-topics-to-inca.ipynb <==
## 06-add-topics-to-inca

**Purpose**: Add topic assignments from multiple topic models into the ES documents which contain the right-wing media outlets' articles.

**Steps**:
1. Create a list of dictionaries where each dict corresponds to a `doc_id`.
2. Each dict contains the `doc_id` and and multiple topic model-related keys.

    - `lda_tfidf_texts_10_top_topic`
    - `lda_tfidf_texts_10_top_topic_pct`
    - `lda_tfidf_texts_10_topic_tokens`
    - `lda_tfidf_texts_10_doc_tokens`
    
    - `lda_tfidf_texts_25_top_topic`
    - `lda_tfidf_texts_25_top_topic_pct`
    - `lda_tfidf_texts_25_topic_tokens`
    - `lda_tfidf_texts_25_doc_tokens`
    
    - `lda_tfidf_texts_40_top_topic`
    - `lda_tfidf_texts_40_top_topic_pct`
    - `lda_tfidf_texts_40_topic_tokens`
    - `lda_tfidf_texts_40_doc_tokens`

3. INCA has already been modified so it can update a document based on a `doc_id` and add multiple new fields in one-go (per [05-softcosine-clusters-inca.ipynb](https://github.com/wlmwng/us-right-media/blob/develop/usrightmedia/code/07-newsevents/)).

import os
import pandas as pd
import copy

from usrightmedia.shared.topics_utils import *

from usrightmedia.shared.loggers import get_logger
LOGGER = get_logger(filename = '06-add-topics-to-inca', logger_type='main')

from inca import Inca
myinca = Inca()

### 1.0 Load `doc_id`s and topic assignments

def load(model):
    df = pd.read_pickle(os.path.join(MODELS_DIR, model, f"{model}_top_topic_with_ids.pkl"))
    # https://stackoverflow.com/a/70311963: add prefix to all columns except doc_id
    df = df.set_index('doc_id').add_prefix(f'lda_tfidf_texts_{model.split("_")[-1]}_').reset_index()
    return df

%%time
df10 = load('lda_corpus_tfidf_docs_texts_topics_10')
df25 = load('lda_corpus_tfidf_docs_texts_topics_25')
df40 = load('lda_corpus_tfidf_docs_texts_topics_40')

df10

#### 1.1 Convert dataframes to dictionaries

%%time
d10 = df10.to_dict("records")
d25 = df25.to_dict("records")
d40 = df40.to_dict("records")

d10[-5:-3]

### 2.0 Prepare inputs for INCA

- Similar steps to [05-softcosine-clusters-inca.ipynb](https://github.com/wlmwng/us-right-media/blob/5156cf1590341828ca00adceeb9b8beebc612fbe/usrightmedia/code/07-newsevents/05-softcosine-clusters-inca.ipynb)

def prep(docs):
    """Prepare docs' input format for INCA.
    Renames 'doc_id' to '_id' so INCA will know which ES document to update.
    
    Args: 
        docs (list of dicts)    
    
    Returns:
        formatted docs (list of dicts)    
    """
    
    # Remove 'doc_id' so it isn't added as a redundant field on the ES document. Note this modifies the original object.
    for doc in docs:
        doc["_id"] = doc["doc_id"]
        doc.pop("doc_id")

%%time
prep(d10)
prep(d25)
prep(d40)

### 3.0 Update Elasticsearch database through INCA

def update(docs):
    # each dict in docs will no longer have an "_id" as its popped off in INCA
    # https://github.com/wlmwng/inca/blob/1c1ed382b2db34682234af0298cf4c19ecbc7074/inca/core/database.py#L209
    myinca.database.update_documents(docs, batchsize=2000)

- check Kibana after running `update()` on `AmericanRenaissance_1128638341`

```
GET /inca_alias/_search
{"_source": {
            "excludes": [ "META" ]
        },
        "query": {
            "bool": {
                "filter": [
                    {"term": {"_id": "AmericanRenaissance_1128638341"}}
                ]
            }
        }
    }
```


%%time
update(d10)

%%time
update(d25)

%%time
update(d40)

d40[11:12]
