==> usrightmedia/code/08-topics/07-topic-model-review.ipynb <==
## 07-topic-model-review

**Purpose**: Evaluate topic models by reviewing samples of docs per topic.

import os
import pandas as pd

from usrightmedia.shared.topics_utils import *

from usrightmedia.shared.loggers import get_logger
LOGGER = get_logger(filename = '07-topic-model-review', logger_type='main')

### 1. Load topic models' summary and documents

def load(model):
    df_smry = pd.read_pickle(os.path.join(MODELS_DIR, model, f"{model}_top_topic_summary.pkl"))
    df_docs = pd.read_pickle(os.path.join(MODELS_DIR, model, f"{model}_top_topic_with_ids.pkl"))
    
    # fix percentages
    return df_smry, df_docs

dfs_10, dfd_10 = load('lda_corpus_tfidf_docs_texts_topics_10')
dfs_25, dfd_25 = load('lda_corpus_tfidf_docs_texts_topics_25')
dfs_40, dfd_40 = load('lda_corpus_tfidf_docs_texts_topics_40')

with pd.option_context('display.max_colwidth', None):
    display(dfs_10)
    display(dfs_25)
    display(dfs_40)

### 

dfd_10
